requirements.txt
numpy==1.24.3
torch==2.2.2
sentence-transformers==2.2.2
huggingface_hub==0.16.4
pdfminer.six